Rename all nms arrow classes
