+++ src/com/cnaude/chairs/vehiclearrow/nmsV17R1/NMSChairsArrow.java
@@ -1,11 +1,11 @@
-package com.cnaude.chairs.vehiclearrow.nms164;
+package com.cnaude.chairs.vehiclearrow.nmsV17R1;
 
-import net.minecraft.server.v1_6_R3.EntityArrow;
+import net.minecraft.server.v1_7_R1.EntityArrow;
 
 import org.bukkit.Location;
 import org.bukkit.Server;
-import org.bukkit.craftbukkit.v1_6_R3.CraftServer;
-import org.bukkit.craftbukkit.v1_6_R3.CraftWorld;
+import org.bukkit.craftbukkit.v1_7_R1.CraftServer;
+import org.bukkit.craftbukkit.v1_7_R1.CraftWorld;
 import org.bukkit.entity.Arrow;
 
 import com.cnaude.chairs.vehiclearrow.NMSChairsArrowInterface;
@@ -17,7 +17,7 @@ public class NMSChairsArrow extends EntityArrow implements NMSChairsArrowInterfa
 	}
 
 	@Override
-	public void l_() {
+	public void h() {
 	}
 
 	@Override
